requirements.txt
asgiref==3.3.0
bigcommerce==0.20.1
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.3
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
install==1.3.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.4
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
